requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
fonttools==4.47.0
idna==3.6
inference-sdk==0.9.8
kiwisolver==1.4.5
marshmallow==3.20.2
matplotlib==3.8.2
mypy-extensions==1.0.0
numpy==1.26.3
opencv-python==4.9.0.80
opencv-python-headless==4.8.1.78
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.4
six==1.16.0
supervision==0.17.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
